Adjust type defs
