+++ tsconfig.json
@@ -25,6 +25,5 @@
     "styles.d.ts",
     "i18n.d.ts",
     "site.d.ts",
-    "pages.d.ts"
   ]
 }
